# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RestKG API",
    "version": "1.0.0",
    "description": "API documentation for RestKG"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/farmers": {
      "get": {
        "summary": "Get all farmers",
        "description": "Retrieve a list of all farmers.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer", "example": 1 },
                      "name": { "type": "string", "example": "Farmer 1" },
                      "description": { "type": "string", "example": "Description for Farmer 1" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new farmer",
        "description": "Add a new farmer to the list.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "New Farmer" },
                  "description": { "type": "string", "example": "Description for New Farmer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Farmer created successfully"
          }
        }
      },
      "put": {
        "summary": "Update a farmer",
        "description": "Update an existing farmer's details.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer", "example": 1 },
                  "name": { "type": "string", "example": "Updated Farmer" },
                  "description": { "type": "string", "example": "Updated description" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Farmer updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a farmer",
        "description": "Remove a farmer from the list.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Farmer deleted successfully"
          }
        }
      }
    },
    "/api/worker": {
      "get": {
        "summary": "Get all workers",
        "description": "Retrieve a list of all workers.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer", "example": 1 },
                      "name": { "type": "string", "example": "Worker 1" },
                      "description": { "type": "string", "example": "Description for Worker 1" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new worker",
        "description": "Add a new worker to the list.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "New Worker" },
                  "description": { "type": "string", "example": "Description for New Worker" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Worker created successfully"
          }
        }
      },
      "put": {
        "summary": "Update a worker",
        "description": "Update an existing worker's details.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer", "example": 1 },
                  "name": { "type": "string", "example": "Updated Worker" },
                  "description": { "type": "string", "example": "Updated description" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Worker updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a worker",
        "description": "Remove a worker from the list.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker deleted successfully"
          }
        }
      }
    }
  }
}